requirements.txt
Flask
FastAPI
SQLAlchemy
psycopg2-binary
python-dotenv
uvicorn
openai

